# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Online Service Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ServiceRequestStatus = typeof ServiceRequestStatus[keyof typeof ServiceRequestStatus];


export const ServiceRequestStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface UploadedDocument {
  id: number;
  requestId: number;
  docType: string;
  filename: string;
  url: string;
}

export interface ServiceRequest {
  id: number;
  fullName: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  city: string;
  serviceType: string;
  /** @nullable */
  notes?: string | null;
  status: ServiceRequestStatus;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  paymentScreenshot?: string | null;
  documents?: UploadedDocument[];
  createdAt: string;
  updatedAt?: string;
}

export interface ServiceRequestInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 10 */
  mobile: string;
  email?: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  serviceType: string;
  notes?: string;
}

export type ServiceRequestUpdateStatus = typeof ServiceRequestUpdateStatus[keyof typeof ServiceRequestUpdateStatus];


export const ServiceRequestUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface ServiceRequestUpdate {
  status?: ServiceRequestUpdateStatus;
  notes?: string;
}

export interface PaymentInfo {
  transactionId?: string;
  paymentScreenshot?: string;
}

export interface DocumentUpload {
  requestId: number;
  docType: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  success: boolean;
  message?: string;
}

export interface AdminStats {
  total: number;
  pending: number;
  processing: number;
  completed: number;
  rejected: number;
  today: number;
}

export type ListRequestsParams = {
status?: string;
service?: string;
mobile?: string;
date?: string;
};