# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Online Service Center API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  total: number;
  pending: number;
  processing: number;
  completed: number;
  rejected: number;
  today: number;
}