# lib/api-zod/src/generated/types/serviceRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Online Service Center API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceRequestStatus } from './serviceRequestStatus';
import type { UploadedDocument } from './uploadedDocument';

export interface ServiceRequest {
  id: number;
  fullName: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  city: string;
  serviceType: string;
  /** @nullable */
  notes?: string | null;
  status: ServiceRequestStatus;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  paymentScreenshot?: string | null;
  documents?: UploadedDocument[];
  createdAt: string;
  updatedAt?: string;
}