# lib/api-zod/src/generated/types/serviceRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Online Service Center API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceRequestInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 10 */
  mobile: string;
  email?: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  serviceType: string;
  notes?: string;
}