# lib/api-zod/src/generated/types/serviceRequestUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Online Service Center API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceRequestUpdateStatus } from './serviceRequestUpdateStatus';

export interface ServiceRequestUpdate {
  status?: ServiceRequestUpdateStatus;
  notes?: string;
}