# lib/api-zod/src/generated/types/uploadedDocument.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shubham Online Service Center API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadedDocument {
  id: number;
  requestId: number;
  docType: string;
  filename: string;
  url: string;
}